browser-tests/pyproject.toml
[project]
name = "verso-slides-tests"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pytest",
    "pytest-playwright",
    "beautifulsoup4",
]
